# Speed up uploading/downloading of file and directories.