--- a/CHANGELOG.D/958.feature
+++ b/CHANGELOG.D/958.feature
@@ -1,0 +1,1 @@
+`neuro storage cp` is now up to 2 times faster for a directory with many small files.