Before optimization:
1 minute for `neuro storage cp -r neuromation storage:`
1 minute for `neuro storage cp -r storage:neuromation /tmp`

After optimization:
36 seconds for `neuro storage cp -r neuromation storage:`
34 seconds for `neuro storage cp -r storage:neuromation /tmp`

For comparison:
From 22 seconds to 1 minute 16 seconds for `neuro storage load -r neuromation storage:`
18 seconds for `neuro storage load -r storage:neuromation /tmp`
7 seconds for `neuro storage cp neuromation.tar storage:`
2 seconds for `neuro storage cp storage:neuromation.tar /tmp`