@dalazx 2019-08-15
@serhiy-storchaka curious, in case of `load`, what's the overhead of spawning a job, provided a local docker image is present already?