@asvetlov 2019-08-15
> what's the overhead of spawning a job, provided a local docker image is present already?

About 20 secs IIRC