@dalazx 2019-08-15
`18 seconds for neuro storage load -r storage:neuromation /tmp`
does that include launching a job?