@serhiy-storchaka 2019-08-15
Yes, it is a total time.